ISSUE-2: Add support for self-managing controllers
Controllers is used for moving sprites around on the canvas. Currently,
it will get the id of the object that it is supposed to control, and
then the Game object will tell it when it is time to start working.

The controller will then itself determine when to do updates (i.e it
will use setTimeout instead of relying on the main loop).